requirements.txt
feedparser==6.0.11
requests==2.31.0
pydantic==2.7.1
beautifulsoup4==4.12.3
streamlit==1.35.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.0
python-dotenv==1.0.1
pytest==8.2.0
pytest-mock==3.14.0
